# A little optimization