--- a/if_ver.vim
+++ b/if_ver.vim
@@ -16,8 +16,8 @@ echo "\nPython 3:"
 python3 import sys; print(sys.version)
 echo "\nRuby:"
 ruby print RUBY_VERSION
-echo "\nTcl:"
-tcl puts [info patchlevel]
+" echo "\nTcl:"
+" tcl puts [info patchlevel]
 echo "\n"
 echo "\n $VIMRUNTIME"
 echo $VIMRUNTIME