… exclist.txt
… gitlog.txt
… vimdll-huge.cfg