Reduced the download size of the installation file of the program Python ver. 3.
Reduced dependency on third-party tools such as cygwin and jq.
Where possible, a linking mechanism has been used instead of copying files.
Other.

Execution time of “install”, “build” and “package” sections without dependency
downloading:
it was
x64 - 631.769sec.
x86 - 657.733sec.
it became
x64 - 328.226sec.
x86 - 335.175sec.