@chrisbra 2025-01-19
can you please rebase? 

What is the purpose of additional patch? Ideally we wouldn't need an extra patch here.

Also please in the future keep the style and don't change it. It makes reviewing harder if you e.g. change `nmake` to `nmake.exe` or `nmake -f` to `nmake -lf`. Or at least use a separate commit for those stylistic chagnes. This makes reviewing much harder otherwise.